fix(command): Move tsconfig.json into src folder
`jest-runner-tsc` was looking for the `tsconfig.json` file at the jest `rootDir` which we want to be the `cwd` but want the `tsconfig.json` to live in the `src` directory.

With #8, learned a way around that for linting so move the workaround into the base jest config so that all can benefit.