chore(release): Set up release build stages (#23)
Leverages Travis build stages. Uses `files` property in `package.json` + `.npmignore` in order to dictate which files should be included. The build step will transpile the unit tests, but they are not published because of the `.npmignore`.

Removed the `.babelrc` from the root so now both the bootstrap and the `test` command have to do some special connection with `babel-config-cjs.js`