Upgrading packages to latest
Needed to bump to `jest`@24 which used `@babel/core` (instead of `babel-core`). Before I was able to use Yarn resolutions to map `babel-core` to `@babel/core`, but an app using `@benmvp/cli` would have to do the same thing, which is less than ideal. This was found while trying to update [`url-lib`](https://github.com/benmvp/url-lib) to use `@benmvp/cli`.

Also needed to bump to latest of `@typescript-eslint/*` so that it will work w/ the latest version of `typescript` we're using and stop giving the warning in `typescript-estree`.

However, bumping some packages, meant bumping others, so in the end, I just updated everything I could to the latest. The biggest update was switching from `eslint-pluging-typescript` (abandoned) to `@typescript-eslint/eslint-plugin`. This came w/ a set of recommended rules as well.

BREAKING CHANGE: Typescript eslint rules have been updated to catch more issues.

Most new lint rules are fixable, but likely there will be code that was "fine" before that will need to be manually fixed.