patch(commands): Delete unwated built files
When running babel we copy over some JS & JSON files we don't need to transpile, but as a result the test files get copied over to built locations which we don't want.

So now after finishing with the transpiling and generation of type declarations, we delete test files (either in `__tests__` folder or with `.spec` extension).