feat(command): add integrate command (#47)
Added a new `integrate` command which supports running integration tests from a separate `integration-tests/` "project" within the library. Docs for the new `integrate` command were added as well.

I also added integration tests for `@benmvp/cli` for the build command to verify that build actually does what it is supposed to. So I was able to Fix #32 and also dogfood the `integrate` command.

As part of building out the `integrate` functionality, I needed to add the result of `npm pack` as a dev dependency to the project, which kept failing in Yarn. `yarn add --dev /path/to/tarball.tgz` would return before it was _really_ done and a result the follow on commands with fail. As a result I switched the project over from Yarn to npm. All docks were updated accordingly.

BREAKING CHANGE: The `unit` test mode was renamed to `spec` in order to make more sense with integration tests.

The way the integration tests are run in the project is by running `npx benmvp test`. The parameters passed to `benmvp integrate` are ultimately passed as the parameters to `npx benmvp test`. Have a mode called `unit` when running integration tests seemed confusion.  Really it's just running Jest spec tests, so the rename was to better reflect that. In order to just run **unit** tests, you run `benmvp test -m spec` in order to just run **integration** tests you run `benmvp integration -m spec`.

I also set `maxWarnings` for eslint to `0` so that all warnings are also errors as well.