fix(command): Install @benmvp/cli w/ integrate (#51)
Integration tests in `url-lib` were failing because a bunch of dependencies were missing. I assumed that I could just run `npx benmvp test` in the integration test to cover everything, but that was only working here because the tarball had those dependencies. Any other repo wouldn't.

So now as part of the integrate process, `@benmvp/cli` gets added as a dev dependency to the integration tests project.