fix(command): Fix install of @benmvp/cli w/ integrate (#52)
The intention with 555a184 was to install `@benmvp/cli` as a dev dependency in the integration tests project, but I just updated the `echo` string now the actual command. 🤦🏾‍♂️

Verified this in `url-lib` by running `npm pack` in this repo and manually installing it in `url-lib` and running its integration tests. 🤞🏾