fix(command): Build before integrate
We need to `build` the library before calling `npm pack` to ensure it has the latest built assets, which will then be used by the integration tests project.

Didn't catch this case when running the integration tests for @benmvp/cli itself because we always `build` before running all of the scripts.