fix(command): Inline Babel helpers with build
When the helpers were not inlined they depended on `@babel/runtime` which was fine for this package, but for the packages uses `@benmvp/cli` they would also depend on `@babel/runtime`, which is not desired.

So now those helpers are just inlined and duplicated. For now if a lib wants to use `async`