feat(lint): Switch to eslint-config-airbnb (#72)
`@benmvp/cli` was using [`eslint-config-eventbrite-react`](https://www.npmjs.com/package/eslint-config-eventbrite-react) for linting, but it's not regularly updated. As a result, greenkeeper kept submitting PRs to update various eslint-related packages, but we couldn't because `eslint-config-eventbrite-react` needed a certain version of `eslint`.

The fix was to switch over to the industry standard [`eslint-config-airbnb`](https://www.npmjs.com/package/eslint-config-airbnb), which is more regularly updated and allows us to upgrade `eslint` to the latest version.

Also:
- Some of the eslint rules are different for the new config, so a bunch of files were fixed to pass the new styling rules. The biggest change was adding the space within object literals
- Switching over to Github actions for CI & CD and removing Travis
- Needed to add `react` & `react-dom` as dev dependencies because `enzyme-adapter-react-16` needs them because it does a `require('react')` inside (for some reason)
- Temporarily removing typescript eslint rules that require type because there were issues around the `tsconfig.json` that I couldn't figure out. this was a result of the upgrade to v2 of `@typescript-eslint/eslint-plugin`
- Removed `jest-environment-enzyme` in favor of including the adapter directly since I hadn't been able to get `enzyme` to passing type checking w/o importing it

BREAKING CHANGE: Some of the eslint rules are different for the new config