feat(lint): Add prettier (#76)
## Problem

Want to add [`prettier`](https://prettier.io/) support mainly to make development easier so that code auto-formats on save. But if we have `prettier,` we might as well integrate it into CI somehow.


## Solution

The solution is multi-pronged:

- Added `.prettierrc.json` and `.vscode/settings.json` in order to have code formatting on save. This applies to all file times `.ts`, `.js`, `.json`, `.md`, etc
- Added a `.github/workflows/format.yml` workflow that runs the `prettier` CLI and overwrites the files with the changes. If there are file changes, it pushes a new auto-commit to the branch with new changes. That way any changes that were missed will get caught by CI. As a result of the push, CI will run _again_ and we can make sure that running `prettier` didn't break any test with malformed files. Since this was the first run of `prettier`, **lots** of files were updated
- Added `prettier` check as part of linting with [`@typescript-eslint/eslint-plugin`](https://github.com/typescript-eslint/typescript-eslint/tree/master/packages/eslint-plugin). This will fail `benmvp test` if there are malformed files, but most should be fixable when running `benmvp start`
- `benmvp integrate` runs `benmvp test` within the `integration-tests/` project, but it needs the same `.prettierrc.json` so those are copied over as part of the script that runs the integration test

In the end, it'll be the responsibility of `benmvp start` to create a new project with prettier configs. They will be something that each library can change themselves, allowing a bit of configuration, but the idea is that they wouldn't be. I had to make the file part of the host library instead of within `benmvp-cli` in order to make the dev environment work properly.

BREAKING CHANGE: Although CI will auto-format, `benmvp test` will fail because prettier now runs on all the files