feat(create): Add create command (#79)
## Problem

Needed to support the `create` command so that a new library with all the infra could be created. Also needed the ability to be able to update an existing library with the infra, either an existing library that had never used `@benmvp/cli` or one that just needs updated infra.

## Solution

Build out the command to:

1. Initialize the repo with git and npm
1. Set up `LICENSE`, `.gitignore`, and `CODE_OF_CONDUCT.md`
1. Set up the `package.json` correctly
1. Copy over the necessary local configuration files

The integration tests are what really validate that a new lib can be created (or updated), but there are unit tests for updating the `package.json`.

The output directory specified in the `create` command gets passed along to the `build` script stored in the repo's `package.json` so we could no longer default it to `lib/`.

In order to copy over the local config files, I had to add a step to the repo's `build` script that copies over the files. This is **not** part of the CLI, but just a build step for the repo. Then the `create` command copies them over locally and replaces some text to reference the new repo's name.

Fixed `spawnAsync` so that the sub-process writes to the standard output just as if it was the parent process (via `{stdio: 'inherit'}`).