fix(create): Add messaging when done (#83)
## Problem

After running the `create` command there's no messaging indicating that the library was successfully created. In fact, the last thing rendered to the screen is the `cp` command that copies over the config files. I actually thought the command was broken and the `cp` had failed.


## Solution

Replaced the `cp` command with friendly text that says "Copying repo config files." and more importantly when the creation is completed it says "[repoName] successfully created" if the library name was specified (new lib) or "Repo successfully updated" for an existing repo.

Also added an update to the `release` workflow to run the unit tests one more time as a double check to make sure we're not releasing broken code that's merged into `master`.